Add files via upload
I was originally going to continue making my 2d game from week#14 but I come up with an idea for horror-like art piece after taking my little cousins to the Minecraft movie. I'm not the biggest fan of Minecraft but I do know about a creature in the game called the endermen. These guys will try to kill you if you look at them so I wanted to make a spinning cube that has an eye following around it but you cant tell its an eye until it rotates at a certain spot where the eye pop ups at you. Its not a big crazy thing but I liked making it and I think it is reminisce of those endermen.